@@ -6,7 +6,7 @@ dependencies {
     implementation 'com.squareup:javapoet:1.10.0'
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_7
-targetCompatibility = JavaVersion.VERSION_1_7
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
 
 apply from: 'publish.gradle'
