@@ -1,6 +1,6 @@
 package com.mindorks.placeholderview.annotations;
 
-import android.support.annotation.LayoutRes;
+import androidx.annotation.LayoutRes;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
